Upgrade wat-lsp to 0.8.0 and add syntax highlighting to instruction code blocks
