+++ exercises/001_hello.wat
@@ -1,11 +1,13 @@
 (;
   We often export and import when writing WASM. 
-  Here we import a function called $log_num that takes in one i32 as its only parameter
-  We import it from the namespace "env" under the name "log" and call it $log_num locally
+  Here we import a function called $log_num that takes in one i32 as its only parameter.
+  We import it from the namespace "env" under the name "log" and call it $log_num locally.
 
-  Calling a function uses the syntax (call $func_name (param1) (param2) ...)
-  Number literals use the form: (TYPE.const NUMBER)
-  e.g. (f64.const 1093.31) or (i32.const -3029)
+  Calling a function uses the syntax `(call $func_name (param1) (param2) ...)`
+  Number literals use the form: `(TYPE.const NUMBER)`
+  e.g. 
+    (f64.const 1093.31)
+    (i32.const -3029)
 
   Call $log_num with the 32-bit integer 42.
 ;)
