+++ web-build/package.json
@@ -19,7 +19,7 @@
     "@codemirror/lint": "^6.8.0",
     "@codemirror/state": "^6.5.0",
     "@codemirror/view": "^6.36.0",
-    "@emnudge/wat-lsp": "^0.7.0",
+    "@emnudge/wat-lsp": "^0.8.0",
     "canvas-confetti": "^1.9.0",
     "codemirror": "^6.0.0",
     "js-wasm-tools": "^1.3.1",
