+++ web-build/src/style.css
@@ -86,6 +86,22 @@ body {
   letter-spacing: -0.01em;
 }
 
+.sidebar-github {
+  color: var(--text-dim);
+  display: flex;
+  align-items: center;
+  padding: 4px;
+  border-radius: 4px;
+  transition:
+    color 0.15s,
+    background 0.15s;
+}
+
+.sidebar-github:hover {
+  color: var(--text);
+  background: rgba(137, 180, 250, 0.1);
+}
+
 .sidebar-toggle {
   background: none;
   border: none;
@@ -187,6 +203,15 @@ body {
   border-radius: 3px;
   font-family: var(--font-mono);
   font-size: 13px;
+}
+
+.instructions a {
+  color: var(--accent);
+  text-decoration: none;
+}
+
+.instructions a:hover {
+  text-decoration: underline;
 }
 
 .instructions pre {
